requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiosqlite==0.20.0
    # via -r requirements.in
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
black==24.4.0
    # via -r requirements.in
click==8.1.7
    # via
    #   black
    #   uvicorn
fastapi==0.110.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.7
    # via anyio
iniconfig==2.0.0
    # via pytest
mako==1.3.3
    # via alembic
markupsafe==2.1.5
    # via mako
mypy-extensions==1.0.0
    # via black
packaging==24.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
pluggy==1.4.0
    # via pytest
pydantic==2.7.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.18.1
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pytest==8.1.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.23.6
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.29
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   aiosqlite
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.29.0
    # via -r requirements.in
